+++ db/postgres.go
@@ -42,200 +42,6 @@ type postgresDB struct {
 	clock clock.Clock
 }
 
-func (db *postgresDB) GetPlayer(ctx context.Context, id string) (*model.Player, error) {
-	p, err := db.getPlayer(ctx, id)
-	if err != nil {
-		if errors.Is(err, pgx.ErrNoRows) {
-			return nil, ErrPlayerNotFound
-		}
-		return nil, err
-	}
-	return p, nil
-}
-
-func (db *postgresDB) SavePlayer(ctx context.Context, p *model.Player) error {
-	old, err := db.getPlayer(ctx, p.ID)
-	if err != nil {
-		if errors.Is(err, pgx.ErrNoRows) {
-			// This is an insert
-			err := db.insertPlayer(ctx, p)
-			if err != nil {
-				return fmt.Errorf("error inserting player: %w", err)
-			}
-			return nil
-		}
-
-		return fmt.Errorf("error reading player at start of SavePlayer(): %w", err)
-	}
-
-	// This is an update, see what, if anything changed
-	changes, err := db.calculateChanges(old, p)
-	if err != nil {
-		return fmt.Errorf("error calculating changes: %w", err)
-	}
-	if len(changes) > 0 {
-		return db.updatePlayer(ctx, p, changes)
-	}
-	return nil
-}
-
-func (db *postgresDB) DeleteNickname(ctx context.Context, id, oldNickname string) error {
-	const query = `UPDATE players SET nickname1=NULL WHERE id=@id`
-	args := pgx.NamedArgs{
-		"id": id,
-	}
-
-	change := &model.Change{
-		Time:         db.clock.Now().UTC(),
-		PropertyName: "Nickname1",
-		OldValue:     oldNickname,
-		NewValue:     "",
-	}
-
-	tx, err := db.pool.Begin(ctx)
-	if err != nil {
-		return err
-	}
-	defer tx.Rollback(ctx)
-
-	if _, err := tx.Exec(ctx, query, args); err != nil {
-		return fmt.Errorf("error deleteing player nickname (%s): %w", id, err)
-	}
-	if err := insertPlayerChange(ctx, tx, id, change); err != nil {
-		return fmt.Errorf("error inserting player change (%s): %w", id, err)
-	}
-
-	return tx.Commit(ctx)
-}
-
-func (db *postgresDB) Search(ctx context.Context, q string, pos model.Position, team *model.NFLTeam) ([]model.Player, error) {
-	const query = `SELECT id, yahoo_id, name_first, name_last, nickname1,
-				  		position, team, weight_lb, height_in, birth_date,
-						rookie_year, years_exp, jersey_num, depth_chart_order,
-						college, active, created, updated
-					FROM players WHERE fts_player @@ websearch_to_tsquery(@q)
-						AND team ILIKE @team
-						AND position ILIKE @pos`
-
-	const teamAndPosQuery = `SELECT id, yahoo_id, name_first, name_last, nickname1,
-					    		position, team, weight_lb, height_in, birth_date,
-					  			rookie_year, years_exp, jersey_num, depth_chart_order,
-					  			college, active, created, updated
-				  			FROM players WHERE team ILIKE @team AND position ILIKE @pos`
-
-	teamQ := "%"
-	if team != nil {
-		teamQ = team.String()
-	}
-	posQ := "%"
-	if pos != model.POS_UNKNOWN {
-		posQ = string(pos)
-	}
-
-	args := pgx.NamedArgs{
-		"q":    q,
-		"team": teamQ,
-		"pos":  posQ,
-	}
-
-	qq := query
-	if q == "" {
-		qq = teamAndPosQuery
-	}
-	rows, err := db.pool.Query(ctx, qq, args)
-	if err != nil {
-		return nil, fmt.Errorf("error running search query: %w", err)
-	}
-
-	results := make([]model.Player, 0, 8)
-	for rows.Next() {
-		p, err := scanPlayer(rows)
-		if err != nil {
-			return nil, err
-		}
-		results = append(results, *p)
-	}
-
-	return results, nil
-}
-
-func (db *postgresDB) SavePlayerScores(ctx context.Context, leagueID int32, week int, scores []model.PlayerScore) error {
-	const insert = `INSERT INTO player_scores(player_id, league_id, week, score) 
-			VALUES (@playerID, @leagueID, @week, @score)`
-
-	tx, err := db.pool.Begin(ctx)
-	if err != nil {
-		return fmt.Errorf("error starting transaction: %w", err)
-	}
-	defer tx.Rollback(ctx)
-
-	for _, s := range scores {
-		args := pgx.NamedArgs{
-			"playerID": s.PlayerID,
-			"leagueID": leagueID,
-			"week":     week,
-			"score":    s.Score,
-		}
-		if _, err := tx.Exec(ctx, insert, args); err != nil {
-			return fmt.Errorf("error inserting score for %s in league %d: %w", s.PlayerID, leagueID, err)
-		}
-	}
-
-	if err := tx.Commit(ctx); err != nil {
-		return fmt.Errorf("error commiting player scores: %w", err)
-	}
-
-	return nil
-}
-
-func (db *postgresDB) GetPlayerScores(ctx context.Context, playerID string) ([]model.SeasonScores, error) {
-	const query = `SELECT p.league_id, p.week, p.score, l.name, l.year FROM player_scores AS p
-            INNER JOIN leagues AS l ON p.league_id=l.id
-			WHERE p.player_id=@playerID ORDER BY p.league_id, p.week`
-
-	rows, err := db.pool.Query(ctx, query, pgx.NamedArgs{"playerID": playerID})
-	if err != nil {
-		return nil, fmt.Errorf("error querying player scores: %w", err)
-	}
-
-	scoreMap := make(map[int32]*model.SeasonScores)
-	for rows.Next() {
-		var leagueID, score int32
-		var week int
-		var leagueName, leagueYear string
-		if err := rows.Scan(&leagueID, &week, &score, &leagueName, &leagueYear); err != nil {
-			return nil, fmt.Errorf("error scanning score: %w", err)
-		}
-
-		if scores, found := scoreMap[leagueID]; found {
-			scores.Scores[week] = score
-		} else {
-			s := &model.SeasonScores{
-				LeagueID:   leagueID,
-				LeagueName: leagueName,
-				LeagueYear: leagueYear,
-				PlayerID:   playerID,
-				Scores:     make([]int32, 19),
-			}
-			s.Scores[week] = score
-			scoreMap[leagueID] = s
-		}
-	}
-	if err := rows.Err(); err != nil {
-		return nil, fmt.Errorf("error with rows: %w", err)
-	}
-
-	results := make([]model.SeasonScores, 0, len(scoreMap))
-	for _, s := range scoreMap {
-		results = append(results, *s)
-	}
-	slices.SortFunc(results, func(a, b model.SeasonScores) int {
-		return strings.Compare(a.LeagueYear, b.LeagueYear)
-	})
-
-	return results, nil
-}
-
 func (db *postgresDB) ListRankings(ctx context.Context) ([]model.Ranking, error) {
 	const query = "SELECT id, ranking_date FROM rankings ORDER BY ranking_date DESC LIMIT 25"
 
@@ -874,61 +680,6 @@ func (db *postgresDB) ListPowerRankings(ctx context.Context, leagueID int32) ([]
 	return results, nil
 }
 
-func (db *postgresDB) ConvertYahooPlayerIDs(ctx context.Context, players []model.YahooPlayer) ([]string, error) {
-	results := make([]string, 0, len(players))
-	for _, p := range players {
-		id, err := db.findByYahooID(ctx, p.YahooID)
-		if errors.Is(err, pgx.ErrTooManyRows) {
-			return nil, fmt.Errorf("multiple results found for yahoo_id: %s", p.YahooID)
-		}
-		if errors.Is(err, pgx.ErrNoRows) {
-			id, err = db.findByPlayerName(ctx, &p)
-		}
-		if err != nil {
-			return nil, err
-		}
-
-		results = append(results, id)
-	}
-
-	return results, nil
-}
-
-func (db *postgresDB) findByYahooID(ctx context.Context, yahooID string) (string, error) {
-	const idQuery = `SELECT id FROM players WHERE yahoo_id=@yahooID`
-
-	rows, err := db.pool.Query(ctx, idQuery, pgx.NamedArgs{"yahooID": yahooID})
-	if err != nil {
-		return "", fmt.Errorf("error querying player with yahoo_id=%s: %w", yahooID, err)
-	}
-
-	return pgx.CollectExactlyOneRow(rows, func(row pgx.CollectableRow) (string, error) {
-		var id string
-		err := row.Scan(&id)
-		return id, err
-	})
-}
-
-func (db *postgresDB) findByPlayerName(ctx context.Context, p *model.YahooPlayer) (string, error) {
-	yahooDetails := fmt.Sprintf("%s - %s %s %v", p.YahooID, p.FirstName, p.LastName, p.Pos)
-
-	fullName := model.TrimNameSuffix(fmt.Sprintf("%s %s", p.FirstName, p.LastName))
-	results, err := db.Search(ctx, fullName, p.Pos, nil)
-	if err != nil {
-		return "", fmt.Errorf("error searching for %s: %w", yahooDetails, err)
-	}
-	if len(results) == 0 {
-		return "", fmt.Errorf("no results found for %s", yahooDetails)
-	}
-	if len(results) > 1 {
-		return "", fmt.Errorf("multiple results found for %s", yahooDetails)
-	}
-
-	f := results[0]
-	log.Printf("match found for yahoo id %s, sleeper id %s - %s %s", yahooDetails, f.ID, f.FirstName, f.LastName)
-	return f.ID, nil
-}
-
 func (db *postgresDB) getPowerRankingTeams(ctx context.Context, pr *model.PowerRanking, leagueID int32) error {
 	const teamQuery = `SELECT 
 				t.team, m.team_name, m.manager_name, t.rank, 
@@ -1039,50 +790,6 @@ func (db *postgresDB) getPlayer(ctx context.Context, id string) (*model.Player, 
 	return result, nil
 }
 
-func scanPlayer(row pgx.Row) (*model.Player, error) {
-	var result model.Player
-	var pos DBPosition
-	var team DBNFLTeam
-	var yahooID, nickname1, college sql.NullString
-	var birthDate, rookieYear pgtype.Date
-	var created, updated pgtype.Timestamptz
-	err := row.Scan(
-		&result.ID,
-		&yahooID,
-		&result.FirstName,
-		&result.LastName,
-		&nickname1,
-		&pos,
-		&team,
-		&result.Weight,
-		&result.Height,
-		&birthDate,
-		&rookieYear,
-		&result.YearsExp,
-		&result.Jersey,
-		&result.DepthChartOrder,
-		&college,
-		&result.Active,
-		&created,
-		&updated)
-
-	if err != nil {
-		return nil, err
-	}
-
-	result.Position = pos.position
-	result.Team = team.team
-	result.YahooID = valueOrEmpty(yahooID)
-	result.Nickname1 = valueOrEmpty(nickname1)
-	result.College = valueOrEmpty(college)
-	result.BirthDate = birthDate.Time
-	result.RookieYear = rookieYear.Time
-	result.Created = created.Time
-	result.Updated = updated.Time
-
-	return &result, nil
-}
-
 func scanRanking(row pgx.Row) (*model.Ranking, error) {
 	var r model.Ranking
 	var date pgtype.Timestamptz
@@ -1097,240 +804,6 @@ func scanRanking(row pgx.Row) (*model.Ranking, error) {
 	r.Date = date.Time.UTC()
 
 	return &r, nil
-}
-
-func (db *postgresDB) getChangesByID(ctx context.Context, id string) ([]model.Change, error) {
-	const query = `SELECT created, prop, old, new FROM player_changes WHERE player=@id ORDER BY created DESC`
-
-	args := pgx.NamedArgs{
-		"id": id,
-	}
-	rows, err := db.pool.Query(ctx, query, args)
-	if err != nil {
-		return nil, err
-	}
-	changes := make([]model.Change, 0, 16)
-	for rows.Next() {
-		var created pgtype.Timestamptz
-		c := model.Change{}
-		err := rows.Scan(&created, &c.PropertyName, &c.OldValue, &c.NewValue)
-		if err != nil {
-			return nil, fmt.Errorf("error scanning player change: %v", err)
-		}
-		c.Time = created.Time
-
-		changes = append(changes, c)
-	}
-
-	return changes, nil
-}
-
-func (db *postgresDB) insertPlayer(ctx context.Context, p *model.Player) error {
-	if p == nil {
-		return errors.New("insertPlayer - player is nil")
-	}
-	const query = `INSERT INTO players (
-		id,
-		yahoo_id,
-		name_first,
-		name_last,
-		nickname1,
-		position,
-		team,
-		weight_lb,
-		height_in,
-		birth_date,
-		rookie_year,
-		years_exp,
-		jersey_num,
-		depth_chart_order,
-		college,
-		active
-	) VALUES (
-		@id,
-		@yahooID,
-		@nameFirst,
-		@nameLast,
-		@nickname1,
-		@position,
-		@team,
-		@weight,
-		@height,
-		@birthDate,
-		@rookieYear,
-		@yearsExp,
-		@jerseyNum,
-		@depthChartOrder,
-		@college,
-		@active
-	)`
-
-	args := namedArgsForPlayer(p, db.clock)
-	_, err := db.pool.Exec(ctx, query, args)
-	if err != nil {
-		return fmt.Errorf("error inserting player(%s): %w", p.ID, err)
-	}
-	return nil
-}
-
-func (db *postgresDB) updatePlayer(ctx context.Context, p *model.Player, changes []model.Change) error {
-	const update = `UPDATE players
-		SET yahoo_id=@yahooID,
-			name_first=@nameFirst,
-			name_last=@nameLast,
-			nickname1=@nickname1,
-			position=@position,
-			team=@team,
-			weight_lb=@weight,
-			height_in=@height,
-			birth_date=@birthDate,
-			rookie_year=@rookieYear,
-			years_exp=@yearsExp,
-			jersey_num=@jerseyNum,
-			depth_chart_order=@depthChartOrder,
-			college=@college,
-			active=@active,
-			updated=@updated
-		WHERE id=@id`
-
-	tx, err := db.pool.Begin(ctx)
-	if err != nil {
-		return err
-	}
-	defer tx.Rollback(ctx)
-
-	args := namedArgsForPlayer(p, db.clock)
-	_, err = tx.Exec(ctx, update, args)
-	if err != nil {
-		return fmt.Errorf("error updating player (%s): %w", p.ID, err)
-	}
-
-	for _, change := range changes {
-		err := insertPlayerChange(ctx, tx, p.ID, &change)
-		if err != nil {
-			return fmt.Errorf("error inserting player change: %w", err)
-		}
-	}
-
-	err = tx.Commit(ctx)
-	if err != nil {
-		return fmt.Errorf("error commiting player transaction: %w", err)
-	}
-
-	p.Changes = append(p.Changes, changes...)
-	slices.SortFunc(p.Changes, func(a, b model.Change) int {
-		return b.Time.Compare(a.Time)
-	})
-
-	return nil
-}
-
-func insertPlayerChange(ctx context.Context, tx pgx.Tx, id string, change *model.Change) error {
-	const insertChange = `INSERT INTO player_changes(
-		player,
-		prop,
-		old,
-		new
-	) VALUES (
-		@playerId,
-		@prop,
-		@old,
-		@new
-	)`
-
-	args := pgx.NamedArgs{
-		"playerId": id,
-		"prop":     change.PropertyName,
-		"old":      change.OldValue,
-		"new":      change.NewValue,
-	}
-	_, err := tx.Exec(ctx, insertChange, args)
-	return err
-}
-
-func (db *postgresDB) calculateChanges(old, new *model.Player) ([]model.Change, error) {
-	changes := make([]model.Change, 0, 1)
-
-	changes = checkChange(changes, db.clock, "YahooId", old.YahooID, new.YahooID)
-	changes = checkChange(changes, db.clock, "FirstName", old.FirstName, new.FirstName)
-	changes = checkChange(changes, db.clock, "LastName", old.LastName, new.LastName)
-	changes = checkChange(changes, db.clock, "Position", string(old.Position), string(new.Position))
-	changes = checkChange(changes, db.clock, "Team", old.Team.String(), new.Team.String())
-	changes = checkChangeInt(changes, db.clock, "Weight", old.Weight, new.Weight)
-	changes = checkChangeInt(changes, db.clock, "Height", old.Height, new.Height)
-	changes = checkChange(changes, db.clock, "BirthDate", old.FormattedBirthDate(), new.FormattedBirthDate())
-	changes = checkChange(changes, db.clock, "RookieYear", old.FormattedRookieYear(), new.FormattedRookieYear())
-	changes = checkChangeInt(changes, db.clock, "YearsExp", old.YearsExp, new.YearsExp)
-	changes = checkChangeInt(changes, db.clock, "Jersey", old.Jersey, new.Jersey)
-	changes = checkChangeInt(changes, db.clock, "DepthChartOrder", old.DepthChartOrder, new.DepthChartOrder)
-	changes = checkChange(changes, db.clock, "College", old.College, new.College)
-	changes = checkChange(changes, db.clock, "Active", fmt.Sprintf("%v", old.Active), fmt.Sprintf("%v", new.Active))
-
-	// Nickname is special because it isn't part of the sleeper data, therefore
-	// only check for a change if the new data sets a nickname.
-	// To delete a nickname use db.DeleteNickname().
-	if new.Nickname1 != "" {
-		changes = checkChange(changes, db.clock, "Nickname1", old.Nickname1, new.Nickname1)
-	}
-	return changes, nil
-}
-
-func checkChange(changes []model.Change, clock clock.Clock, prop, old, new string) []model.Change {
-	if old != new {
-		c := model.Change{
-			Time:         clock.Now().UTC(),
-			PropertyName: prop,
-			OldValue:     old,
-			NewValue:     new,
-		}
-		changes = append(changes, c)
-	}
-	return changes
-}
-
-func checkChangeInt(changes []model.Change, clock clock.Clock, prop string, old, new int) []model.Change {
-	return checkChange(changes, clock, prop, fmt.Sprintf("%d", old), fmt.Sprintf("%d", new))
-}
-
-func namedArgsForPlayer(p *model.Player, clock clock.Clock) pgx.NamedArgs {
-	return pgx.NamedArgs{
-		"id": p.ID,
-		"yahooID": sql.NullString{
-			String: p.YahooID,
-			Valid:  p.YahooID != "",
-		},
-		"nameFirst": p.FirstName,
-		"nameLast":  p.LastName,
-		"nickname1": sql.NullString{
-			String: p.Nickname1,
-			Valid:  p.Nickname1 != "",
-		},
-		"position": &DBPosition{position: p.Position},
-		"team":     &DBNFLTeam{team: p.Team},
-		"weight":   p.Weight,
-		"height":   p.Height,
-		"birthDate": pgtype.Date{
-			Time:  p.BirthDate,
-			Valid: !p.BirthDate.IsZero(),
-		},
-		"rookieYear": pgtype.Date{
-			Time:  p.RookieYear,
-			Valid: !p.RookieYear.IsZero(),
-		},
-		"yearsExp":        p.YearsExp,
-		"jerseyNum":       p.Jersey,
-		"depthChartOrder": p.DepthChartOrder,
-		"college": sql.NullString{
-			String: p.College,
-			Valid:  p.College != "",
-		},
-		"active": p.Active,
-		"updated": pgtype.Timestamptz{
-			Time:             clock.Now().UTC(),
-			InfinityModifier: pgtype.Finite,
-			Valid:            true,
-		},
-	}
 }
 
 func valueOrEmpty(v sql.NullString) string {
