+++ db/postgres_test.go
@@ -52,172 +52,6 @@ func TestMain(m *testing.M) {
 	os.Exit(code)
 }
 
-func TestDB_saveAndLoad(t *testing.T) {
-	ctx := context.Background()
-	p := getPlayer()
-
-	err := testDB.SavePlayer(ctx, p)
-	assertFatalf(t, err == nil, "error saving player: %v", err)
-
-	res, err := testDB.GetPlayer(ctx, p.ID)
-	assertFatalf(t, err == nil, "error retreiving player: %v", err)
-
-	// Make sure that the after saving and retreiving the player, all the fields
-	// are the same.
-	assertEquals(t, "ID", p.ID, res.ID)
-	assertEquals(t, "YahooID", p.YahooID, res.YahooID)
-	assertEquals(t, "FirstName", p.FirstName, res.FirstName)
-	assertEquals(t, "LastName", p.LastName, res.LastName)
-	assertEquals(t, "Nickname1", p.Nickname1, res.Nickname1)
-	assertEquals(t, "Position", p.Position, res.Position)
-	assertEquals(t, "Team", p.Team, res.Team)
-	assertEquals(t, "Weight", p.Weight, res.Weight)
-	assertEquals(t, "Height", p.Height, res.Height)
-	assertEquals(t, "BirthDate", p.BirthDate, res.BirthDate)
-	assertEquals(t, "RookieYear", p.RookieYear, res.RookieYear)
-	assertEquals(t, "YearsExp", p.YearsExp, res.YearsExp)
-	assertEquals(t, "Jersey", p.Jersey, res.Jersey)
-	assertEquals(t, "DepthChartOrder", p.DepthChartOrder, res.DepthChartOrder)
-	assertEquals(t, "College", p.College, res.College)
-	assertEquals(t, "Active", p.Active, res.Active)
-	assertEquals(t, "player changes", 0, len(res.Changes))
-
-	// The originals should not have their created or updated times set.
-	if !p.Created.IsZero() {
-		t.Errorf("expected created time to be zero")
-	}
-	if !p.Updated.IsZero() {
-		t.Errorf("expected updated time to be zero")
-	}
-
-	// The result should have a created time, but not an updated time.
-	if res.Created.IsZero() {
-		t.Errorf("expected res created time to not be zero")
-	}
-	if !res.Updated.IsZero() {
-		t.Errorf("expected res updated time to be zero")
-	}
-
-	// Now update a field and make sure it persists as expected.
-	p.Weight = p.Weight - 5
-	err = testDB.SavePlayer(ctx, p)
-	assertFatalf(t, err == nil, "error saving player after update: %v", err)
-
-	res2, err := testDB.GetPlayer(ctx, p.ID)
-	assertFatalf(t, err == nil, "error getting updated player: %v", err)
-
-	assertEquals(t, "Weight", p.Weight, res2.Weight)
-	assertEquals(t, "Changes", 1, len(p.Changes))
-	// Now updated should not be zero
-	if res2.Updated.IsZero() {
-		t.Errorf("expected res2 updated time to not be zero")
-	}
-
-	// Lookup a player that doesn't exist
-	res3, err := testDB.GetPlayer(ctx, "1111")
-	assertFatalf(t, err != nil, "should have had an error searching for player")
-	assertEquals(t, "error type", true, errors.Is(err, ErrPlayerNotFound))
-	if res3 != nil {
-		t.Errorf("expected res3 to be nil, but was %v", res3)
-	}
-}
-
-func TestDB_search(t *testing.T) {
-	ctx := context.Background()
-
-	// Change the player name since default player returned by getPlayer is used in several places
-	// and may be in the DB multiple times.
-	p := getPlayer()
-	p.ID = "9998" // Set a static ID since we only ever want one player with this name in the DB
-	p.FirstName = "DK"
-	p.LastName = "Metcalf"
-	p.Nickname1 = ""
-
-	err := testDB.SavePlayer(ctx, p)
-	assertFatalf(t, err == nil, "error saving player: %v", err)
-
-	players, err := testDB.Search(ctx, "Metcalf", model.POS_UNKNOWN, nil)
-	assertFatalf(t, err == nil, "error searching for player: %v", err)
-	assertEquals(t, "num players found", 1, len(players))
-
-	players, err = testDB.Search(ctx, "Frank", model.POS_UNKNOWN, nil)
-	assertFatalf(t, err == nil, "error searching for players: %v", err)
-	assertEquals(t, "num players found when searching for Frank", 0, len(players))
-
-	// TODO: add tests for searching by position and team
-}
-
-func TestNicknames(t *testing.T) {
-	ctx := context.Background()
-	p := getPlayer()
-	p.Nickname1 = "" // Make sure no nickname to start
-
-	err := testDB.SavePlayer(ctx, p)
-	assertFatalf(t, err == nil, "error saving player: %v", err)
-
-	p1, err := testDB.GetPlayer(ctx, p.ID)
-	assertFatalf(t, err == nil, "error fetching player: %v", err)
-	assertEquals(t, "Nickname1", "", p1.Nickname1)
-	if len(p1.Changes) != 0 {
-		t.Errorf("should be 0 changes, but instead there are %d", len(p1.Changes))
-	}
-
-	p1.Nickname1 = "nickname"
-	err = testDB.SavePlayer(ctx, p1)
-	assertFatalf(t, err == nil, "error saving player: %v", err)
-
-	// Verify the nickname has been saved
-	p2, err := testDB.GetPlayer(ctx, p.ID)
-	assertFatalf(t, err == nil, "error fetching player: %v", err)
-	assertEquals(t, "Nickname1", "nickname", p2.Nickname1)
-	if len(p2.Changes) != 1 {
-		t.Errorf("should be 1 changes, but instead there are %d", len(p2.Changes))
-	}
-	assertPlayerChange(t, "change[0]", "Nickname1", "", "nickname", &p2.Changes[0])
-
-	// Update the nickname to a new value
-	p2.Nickname1 = "updated nickname"
-	err = testDB.SavePlayer(ctx, p2)
-	assertFatalf(t, err == nil, "error saving player: %v", err)
-
-	// Verify the nickname has been updated and saved correctly
-	p3, err := testDB.GetPlayer(ctx, p.ID)
-	assertFatalf(t, err == nil, "error fetching player: %v", err)
-	assertEquals(t, "Nickname1", "updated nickname", p3.Nickname1)
-	if len(p3.Changes) != 2 {
-		t.Errorf("should be 2 changes, but instead there are %d", len(p3.Changes))
-	}
-	assertPlayerChange(t, "change[0]", "Nickname1", "nickname", "updated nickname", &p3.Changes[0])
-	assertPlayerChange(t, "change[1]", "Nickname1", "", "nickname", &p3.Changes[1])
-
-	// Save the player with no nickname to make sure it isn't accidently deleted
-	// This simulates getting an update from sleeper that doesn't contain the nickname.
-	pNoNick := getPlayer()
-	pNoNick.Nickname1 = ""
-	err = testDB.SavePlayer(ctx, pNoNick)
-	assertFatalf(t, err == nil, "error saving player: %v", err)
-	pAfterUpdate, err := testDB.GetPlayer(ctx, p.ID)
-	assertFatalf(t, err == nil, "error fetching player: %v", err)
-	if !reflect.DeepEqual(p3, pAfterUpdate) {
-		t.Fatalf("players are not equal after saving an empty nickname")
-	}
-
-	// Now delete the nickname
-	err = testDB.DeleteNickname(ctx, p.ID, p3.Nickname1)
-	assertFatalf(t, err == nil, "error deleting player nickname")
-
-	// Verify the nickname has been deleted
-	p4, err := testDB.GetPlayer(ctx, p.ID)
-	assertFatalf(t, err == nil, "error fetching player: %v", err)
-	assertEquals(t, "Nickname1", "", p4.Nickname1)
-	if len(p4.Changes) != 3 {
-		t.Errorf("should be 3 changes, but instead there are %d", len(p4.Changes))
-	}
-	assertPlayerChange(t, "change[0]", "Nickname1", "updated nickname", "", &p4.Changes[0])
-	assertPlayerChange(t, "change[1]", "Nickname1", "nickname", "updated nickname", &p4.Changes[1])
-	assertPlayerChange(t, "change[2]", "Nickname1", "", "nickname", &p4.Changes[2])
-}
-
 // Test all of the various ranking operations, adding, getting, listing, and deleting
 func TestRankings(t *testing.T) {
 	p1 := getPlayerWithName("Justin", "Jefferson")
@@ -991,79 +825,6 @@ func TestPowerRankings_leagueWithNoRankings(t *testing.T) {
 	}
 }
 
-func TestConvertYahooPlayerIDs(t *testing.T) {
-	ctx := context.Background()
-
-	// Insert several players with specific IDs so that they are always consistent and we can count on
-	// unique results when searching for them. These are also all players found in dleeperdata/players.json file
-	players := []model.Player{
-		{ID: "1166", YahooID: "", FirstName: "Kirk", LastName: "Cousins", Position: model.POS_QB, Team: model.TEAM_ATL},
-		{ID: "1339", YahooID: "26658", FirstName: "Zach", LastName: "Ertz", Position: model.POS_TE, Team: model.TEAM_WAS},
-		{ID: "1352", YahooID: "26664", FirstName: "Robert", LastName: "Woods", Position: model.POS_WR, Team: model.TEAM_HOU},
-		{ID: "1992", YahooID: "27589", FirstName: "Allen", LastName: "Robinson", Position: model.POS_WR, Team: model.TEAM_DET},
-		{ID: "2216", YahooID: "", FirstName: "Mike", LastName: "Evans", Position: model.POS_WR, Team: model.TEAM_TBB},
-		{ID: "2359", YahooID: "", FirstName: "Ameer", LastName: "Abdullah", Position: model.POS_RB, Team: model.TEAM_LVR},
-		{ID: "3225", YahooID: "29288", FirstName: "Tyler", LastName: "Boyd", Position: model.POS_WR, Team: model.TEAM_TEN},
-		{ID: "4080", YahooID: "", FirstName: "Zay", LastName: "Jones", Position: model.POS_WR, Team: model.TEAM_ARI},
-		{ID: "4993", YahooID: "31012", FirstName: "Mike", LastName: "Gesicki", Position: model.POS_TE, Team: model.TEAM_CIN},
-		{ID: "7601", YahooID: "", FirstName: "Rondale", LastName: "Moore", Position: model.POS_WR, Team: model.TEAM_ATL},
-		{ID: "8154", YahooID: "", FirstName: "Brian", LastName: "Robinson", Position: model.POS_RB, Team: model.TEAM_WAS},
-		{ID: "10219", YahooID: "", FirstName: "Chris", LastName: "Rodriguez", Position: model.POS_RB, Team: model.TEAM_WAS},
-		{ID: "SEA", YahooID: "", FirstName: "Seattle", Position: model.POS_DEF, Team: model.TEAM_SEA},
-
-		// These players have duplicate yahoo ids and are not used in the main test
-		{ID: "10225", YahooID: "40073", FirstName: "Jonathan", LastName: "Mingo", Position: model.POS_WR, Team: model.TEAM_CAR},
-		{ID: "10444", YahooID: "40073", FirstName: "Cedric", LastName: "Tillman", Position: model.POS_WR, Team: model.TEAM_CLE},
-	}
-
-	for _, p := range players {
-		if err := testDB.SavePlayer(ctx, &p); err != nil {
-			t.Fatalf("error saving player: %v", err)
-		}
-	}
-
-	input := []model.YahooPlayer{
-		{YahooID: "25812", FirstName: "Kirk", LastName: "Cousins", Pos: model.POS_QB},
-		{YahooID: "26658", FirstName: "Zach", LastName: "Ertz", Pos: model.POS_TE},
-		{YahooID: "26664", FirstName: "Robert", LastName: "Woods", Pos: model.POS_WR},
-		{YahooID: "27589", FirstName: "Allen", LastName: "Robinson", Pos: model.POS_WR},
-		{YahooID: "27535", FirstName: "Mike", LastName: "Evans", Pos: model.POS_WR},
-		{YahooID: "28442", FirstName: "Ameer", LastName: "Abdullah", Pos: model.POS_RB},
-		{YahooID: "29288", FirstName: "Tyler", LastName: "Boyd", Pos: model.POS_WR},
-		{YahooID: "30150", FirstName: "Zay", LastName: "Jones", Pos: model.POS_WR},
-		{YahooID: "31012", FirstName: "Mike", LastName: "Gesicki", Pos: model.POS_TE},
-		{YahooID: "33437", FirstName: "Rondale", LastName: "Moore", Pos: model.POS_WR},
-		{YahooID: "34054", FirstName: "Brian", LastName: "Robinson", Pos: model.POS_RB},
-		{YahooID: "40231", FirstName: "Chris", LastName: "Rodriguez", Pos: model.POS_RB},
-		{YahooID: "100026", FirstName: "Seattle", Pos: model.POS_DEF},
-	}
-
-	expected := []string{"1166", "1339", "1352", "1992", "2216", "2359", "3225", "4080", "4993", "7601", "8154", "10219", "SEA"}
-
-	results, err := testDB.ConvertYahooPlayerIDs(ctx, input)
-	if err != nil {
-		t.Fatalf("error converting yahoo player ids: %v", err)
-	}
-	if !reflect.DeepEqual(expected, results) {
-		t.Errorf("expected: %v, got: %v", expected, results)
-	}
-
-	// Search for a player that won't be found
-	notFound := []model.YahooPlayer{
-		{YahooID: "999999", FirstName: "Captain", LastName: "America", Pos: model.POS_RB},
-	}
-	if _, err := testDB.ConvertYahooPlayerIDs(ctx, notFound); err == nil {
-		t.Errorf("expected an error but there wasn't one when looking up Captain America")
-	}
-
-	multipleFound := []model.YahooPlayer{
-		{YahooID: "40073", FirstName: "Cedric", LastName: "Tillman", Pos: model.POS_WR},
-	}
-	if _, err := testDB.ConvertYahooPlayerIDs(ctx, multipleFound); err == nil {
-		t.Errorf("expected an error but there wasn't one when looking up a duplicated ID")
-	}
-}
-
 func getPlayer() *model.Player {
 	id := atomic.AddInt32(&idCtr, 1)
 
@@ -1143,17 +904,5 @@ func assertError(t *testing.T, tcName string, e1, e2 error) {
 	}
 	if e1.Error() != e2.Error() {
 		t.Errorf("errors are not equal in %s, expected: %v, got: %v", tcName, e1, e2)
-	}
-}
-
-func assertPlayerChange(t *testing.T, key, exProp, exOld, exNew string, c *model.Change) {
-	if exProp != c.PropertyName {
-		t.Errorf("%s.PropertyName - expected: '%s', got: '%s'", key, exProp, c.PropertyName)
-	}
-	if exOld != c.OldValue {
-		t.Errorf("%s.OldValue - expected: '%s', got: '%s'", key, exOld, c.OldValue)
-	}
-	if exNew != c.NewValue {
-		t.Errorf("%s.NewValue - expected: '%s', got: '%s'", key, exNew, c.NewValue)
 	}
 }
